Move quaternion utils to nav_utils
