+++ include/turtle_nav_cpp/eigen_utils.hpp
@@ -70,26 +70,6 @@ std::array<double, row * col> MatrixToStdArray(
  */
 const Matrix2d GetCholeskyLower(const Matrix2d & matrix);
 
-/**
- * @brief Convert heading to Eigen quaternion
- *
- * @tparam T
- * @param[in] heading
- * @return Eigen::Quaternion<T>
- */
-template <typename T>
-Eigen::Quaternion<T> HeadingToQuaternion(T heading)
-{
-  return Eigen::Quaternion<T>(Eigen::AngleAxis<T>(heading, Eigen::Matrix<T, 3, 1>::UnitZ()));
-}
-
-template <typename T>
-T QuaternionToHeading(const Eigen::Quaternion<T> & q)
-{
-  // For planar rotation, the quaternion should have the form: (0, 0, cos(th/2)) + sin(th/2)
-  return atan2(q.z(), q.w()) * 2;
-}
-
 }  // namespace eigen_utils
 
 #endif  // TURTLE_NAV_CPP_EIGEN_UTILS_HPP_
